requirements.txt
fastapi[all]
uvicorn
SQLAlchemy
psycopg2-binary
python-dotenv
pydantic[email]
pytest
httpx
pytest-mock
pytest-playwright
python-jose[cryptography]
passlib[bcrypt]
python-multipart